backend/requirements.txt
# FastAPI & Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# CrewAI & AI Agents
crewai==0.28.8
crewai-tools==0.2.6

# LLM & APIs
openai>=1.13.3,<2.0.0  # Compatible with crewai 0.28.8
anthropic>=0.7.1

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
reportlab>=4.0.7

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.1,<3.0.0  # Compatible with crewai-tools 0.2.6
pydantic-settings>=2.1.0
httpx>=0.25.2
setuptools>=65.0.0  # For pkg_resources

# Data & Processing
pandas>=2.1.4
numpy>=1.26.2

# CORS & Middleware
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
